Moves Offerings
